requirements.txt
langchain
langchain-openai
langchain-community
openai
faiss-cpu
pandas
python-dotenv
sentence-transformers
langchain
langchain-community
pandas

